Inpemented Account as abstract class
